requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52

gigachat==0.1.28

pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0

pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

tiktoken>=0.5.2
pytest>=7.4.3
reportlab>=4.0.7